Remove MSE support.
Too much work to keep maintaining it forward.
