+++ lib/playback/context.ts
@@ -1,6 +1,6 @@
 /// <reference types="vite/client" />
 
-import * as Message from "./message"
+import * as Message from "./worker/message"
 
 // This is a non-standard way of importing worklet/workers.
 // Unfortunately, it's the only option because of a Vite bug: https://github.com/vitejs/vite/issues/11823
