+++ lib/playback/tsconfig.json
@@ -1,7 +1,7 @@
 {
 	"extends": "../tsconfig.json",
 	"include": ["."],
-	"exclude": ["./webcodecs/worklet"],
+	"exclude": ["worklet"],
 	"compilerOptions": {
 		"types": ["dom-mediacapture-transform", "dom-webcodecs"]
 	},
