+++ lib/playback/worklet/index.ts
@@ -1,5 +1,5 @@
 // TODO add support for @/ to avoid relative imports
-import { Ring } from "../../../common/ring"
+import { Ring } from "../../common/ring"
 import * as Message from "./message"
 
 class Renderer extends AudioWorkletProcessor {
